12-13 read example data into Vector of tasks, where each task task is a Vector of lines, each of which is a Vector of Vectors of Chars
Issues: Splitting is awkward because we merge the lines into a string with mkString("\n"), that is, mkString throws away the newline, so we write it back in explicitly. Can we use the newlines in the original input instead of having to write in new ones?

Wrapping using() around the Source() function gives funky behavior. Troubleshoot this eventually.